ISSUE-81 Abstract the message handling layer - Add MatcherLayer
