+++ src/main/java/net/mossol/bot/model/LocationInfo.java
@@ -3,7 +3,7 @@ package net.mossol.bot.model;
 import lombok.Data;
 
 @Data
-public class MenuInfo {
+public class LocationInfo {
     private final String title;
     private final double latitude;
     private final double longitude;
