+++ src/main/java/net/mossol/bot/service/MatcherService.java
@@ -1,0 +1,7 @@
+package net.mossol.bot.service;
+
+import net.mossol.bot.model.ReplyMessage;
+
+public interface MatcherService {
+    ReplyMessage match(String requestMessage) throws Exception;
+}
